refactor!: flatten package structure and enforce package-private access
